pass params and route to shared component
